Prevent sending all when not logged in
Sending single samples checks to ensure user is logged in before
sending. The send all option did not do this, allowing creation
of anonymous records. Fixes this problem in the same style as
for single records. When not logged in samples  are set to send
and synced when next logged in.