Fix geolocation timeout on iOS
There was a race condition where the next location request was already
in progress when the stop command was issued. The callback on this
latent request still fired, initiating the timer again. Further attempts
to cancel the timeout failed as its ID had been deleted. By saving the
ID we can have a second attempt at cancelling the timeout which succeeds
as the first attempt did halt the GPS.